% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draculR_counts.R
\name{geometric.mean}
\alias{geometric.mean}
\title{The geometric mean of vector}
\usage{
geometric.mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector with values}

\item{na.rm}{Boolean to remove NAs}
}
\value{
A numeric value
}
\description{
This function has been extracted from psych v2.4.6.26
in order to reduce the amount of libraries required.
It calculates the geometric mean of a given vector.
}
\examples{
geometric.mean(c(2, 3, 4, NA))
geometric.mean(c(2, 3, 4, NA), na.rm=TRUE)
}
